Finished the vertices lookup table
The lookup table is expressed as an FSM written in vhdl that output the
correct value depending on the state it is. "Left" and "right" input change
the set of verticies that is being outputed thus the frame, which will
creat an ilussion of rotation in a room. C++ programs are used to
convert the csv file with the sets of verticies to the corresponding
sections of the vhdl code. To compile the code just run the compile
scripts, and then execute a.out. Except for the states_generator.cpp,
those files will output the terminal. states_generator.cpp outputs to a
.vhdl file because there's too many lines of code for the terminal
buffer.