pull fpga top into dev branch to put everything together (#6)
* New Readme

Added new readme with short description of the project, relevant tool
chains, and

* New readme created

It includes outline of the project, toolchains, and guidelines to
writing the code and the documentation.

* Aesthetic changes

Changed a dumbly typed numbers to actuall html numbered list

* More guidelines

Added guideline point about writing commits, reminder to pull before
push, and what's allowed to be in the repo.

* VGA configuration (.ucf) files

* VGA hardware

* Finished the vertices lookup table

The lookup table is expressed as an FSM written in vhdl that output the
correct value depending on the state it is. "Left" and "right" input change
the set of verticies that is being outputed thus the frame, which will
creat an ilussion of rotation in a room. C++ programs are used to
convert the csv file with the sets of verticies to the corresponding
sections of the vhdl code. To compile the code just run the compile
scripts, and then execute a.out. Except for the states_generator.cpp,
those files will output the terminal. states_generator.cpp outputs to a
.vhdl file because there's too many lines of code for the terminal
buffer.

* Added framebuffers

new framebuffer system allows to write to the screen as if writing to memory instead of having to rely on timing of math.

* Simple testbench for simulating the verticies component

* Moved everything to the src dirrectory

* Created top entity branch.

Right now it only list of vertices

* Moved verices-related fpga hardware to  directory

* Moved files to directory

* Debounce entity + corresponding testbench

* Removed useless debug output from debouncer entity

* Fixed path to source files in test build file

* Added top-level description for FPGA

* Test buildfile for toplevel entity

* Made a test bench for the top entity thing

* Fixed some bugs

* Changed the clock decalration for testing, cause idk what's broken.

* Changed the file names to match the entity names and some small changes.

* Fixed the lack of ready signal in the tb for the top entity.

* Properly added ready signal now.

* Chnaged the of the entities in the testbench file.

It doesn't break the GoWithTheFlow now.

* Fixed rdy signal

* Fixed a typo

* Fixed a missing signal

* Added clk to the sensitivity list

* Fixed a typo in the entity name

* Added input buffer register as a file and to the top entity.

* Added pipeline component and changed some typos in the top entity.

* Added .gitignore

* Added comments and fixed some formatting

* Fixed path in Makefile target

* Finished description of toplevel

* Added memory controller implementation as well as its testbench

* Extra info on the memory controller implementation

* removed maskselect, changed output assignments of bottom process to avoid synthesis errors

Changes were tested on DEV board, wtriting -> storing -> reading arbitrary data at various memory adresses works on human timescale using on-board switches to drive inputs, and leds to inspect outputs

* Progress on the ROP part that connects vga output with the SRAM

Finished MUX and the registers, started working on the main controller.

* Progress on the controller

Port mapping done; states declaration done; some first states done;
signals done; and stufff ya know

* Progress on the controller

New states according to the diagram and correct assignment of the
signals.

* NEW STATES!!!

Not the united one, they didn't capture anything, but states of the
controller fsm. Basically finished the main branch (the one that reads
from teh chip). Next commit will be the one that clears the buffer.

* Finished fsm.

Finished the vga controller fsm, also changed the name of the directory,
so it makes more sense.

* Debugged the stuff (mostly typos) and started the top entity.

* Progress on the top entity added the color buffer and misc changes.

* Added buttons input to the eof_detector; finished the rop.

Top entity now has all the components and all the signals assigned.

* Finished and debugged the rop, also added a compilation script.

* Orginized stuff a bit, got rid of unnecessary old stuff, and changed
names.

* Moved to vga directory, it's a vga subsystem component after all.

* Update memctrl_syncpulses.vhdl

Fixed some comments which used # instead of --. Also added a typecast from unsigned to a logic vector on line 125.

* Fixed top level and made it compatible with the new vga rop component.

* Merged from the list of vertices branch

* Updated VGA_syncpules to work with a 6 Mhz clock divider instead of counting double from a 25 Mhz clock

* Merged from the fpga/vertices branch

* Merged from the fpga/vertices branch

* Added clk didvider

Simple component that makes a clk divided by 8

* Testing syncpulses

* Typo

* Changed a signal to unsigned

* Testing on FPGA

* More color

Added color soutput

* fixed process statement

* Caseted shit to unsigned

whatever

* WIKTOR

* Added blanking color

* Finished VGA Controll system

clk_6 at 6 MHZ, 320x240 resolution

* Added the 6Mhz clock divider and its port mapping

* Remove unnecessary signal

* Adjusted everything for 17 bit vga address

* Adjusted the rop to now use 17 bit addresses and buffers for 320x240

* Added pinassignments and moved unessecary files
moved the files to test folder

* Fixed a missing signal

* Fixed pinassignment

* Fixed incorrectly assigned pins

* Inverting the reset signal

* Merge from the fpga/vertices.

* Fixed a latch

* Added others state

* Fixed sensitivity lists

* Fixed latch and sensitivity list

* Fixed sensitivity list

* Fixed sensitivity list

* Fixed latches and undefined behaviour.

* Update sram.vhdl

Removed compilation warnings

* fixing some sensitivity lists

* FOR SOME REASON SYNCPULSES IS BROKEN

* Fuck VGA....

* AAAAAAAAAAAAAAAA

* It's a vector f u

* Adding more possible sync values

* Added new clk divider and changed the naming scheme.

* Changed the screen address to interpolate

* Changes entity names in the clk divider and added the clk divider in the
top entity

* Fixed this bs

* Finished Syncpulses I hope

* WIKTOR

* Tb for those async buffers

* Fixed the tb

* Changed the name because of questa

* Fixed it now the clk is even

* tb

* Merge from the fpga/vertices

* Actual merege from the last commit

* Removed misplaced file

* Remove the pipelines controller and the register

This was here only for testing of the entire system now it's unnecessary
not synchronized nonsense.

---------

Co-authored-by: Xses <wiktort52@gmail.com>
Co-authored-by: xses <witort52@gmail.com>
Co-authored-by: Thijs van Esch <me@thijsvanesch.com>
Co-authored-by: GyanVDZ <127389477+GyanVDZ@users.noreply.github.com>